UserとCategoryの関連付け、それに伴うデータ取得方法の変更

+++ backend/app/controllers/templates_controller.rb
@@ -1,6 +1,8 @@
 class TemplatesController < ApplicationController
+  before_action :authenticate_api_v1_user!
+
   def index
-    @templates = Template.all
+    @templates = current_api_v1_user.templates
     render json: @templates
   end
 
@@ -10,19 +12,24 @@ class TemplatesController < ApplicationController
   end
 
   def update
-    template = Template.find(params[:id])
+    template = current_api_v1_user.templates.find(params[:id])
     template.update(template_params)
     head :ok
   end
 
   def destroy
-    template = Template.find(params[:id])
-    template.destroy
-    head :ok
+    template = current_api_v1_user.templates.find(params[:id])
+    if template.nil?
+      head :not_found
+    else
+      template.destroy
+      head :ok
+    end
   end
 
   private
   def template_params
-    params.permit(:title, :body, :genre_id)
+    params.require(:template).permit(:title, :body, :category_id).merge(user_id: current_api_v1_user.id)
   end
+
 end
